:building_construction: Move 'Card' to its own component
